+++ src/mse_cli/home/command/code_provider/encrypt.py
@@ -1,4 +1,4 @@
-"""mse_cli.home.command.code_provider.decrypt module."""
+"""mse_cli.home.command.code_provider.encrypt module."""
 
 import sys
 from pathlib import Path
@@ -11,19 +11,20 @@ from mse_cli.log import LOGGER as LOG
 def add_subparser(subparsers):
     """Define the subcommand."""
     parser = subparsers.add_parser(
-        "decrypt", help="decrypt a file using Fernet symmetric encryption"
+        "encrypt", help="encrypt a file using Fernet symmetric encryption"
     )
 
     parser.add_argument(
         "--input",
         type=Path,
         required=True,
-        help="path to the file to decrypt",
+        help="path to the file to encrypt",
     )
 
     parser.add_argument(
         "--key",
         type=Path,
+        metavar="FILE",
         required=True,
         help="path to the file within a 32 bytes key URL Safe Base64 encoded",
     )
@@ -32,7 +33,7 @@ def add_subparser(subparsers):
         "--output",
         type=Path,
         metavar="FILE",
-        help="path to write decrypted file",
+        help="path to write encrypted file",
     )
 
     parser.set_defaults(func=run)
@@ -40,18 +41,18 @@ def add_subparser(subparsers):
 
 def run(args) -> None:
     """Run the subcommand."""
-    LOG.info("Decrypting %s...", args.input)
+    LOG.info("Encrypting %s...", args.input)
 
     key: bytes = args.key.read_bytes()
-    encrypted_data: bytes = args.input.read_bytes()
+    data: bytes = args.input.read_bytes()
 
-    data: bytes = Fernet(key).decrypt(encrypted_data)
+    encrypted_data: bytes = Fernet(key).encrypt(data)
 
     if args.output:
-        args.output.write_bytes(data)
-        LOG.info("File sucessfully decrypted to %s", args.output)
+        args.output.write_bytes(encrypted_data)
+        LOG.info("File encrypted to %s", args.output)
     else:
-        LOG.info("Data sucessfully decrypted!")
+        LOG.info("Data sucessfully encrypted!")
         LOG.info(
             "----------------------------------------------------------------------"
         )