+++ src/mse_cli/home/command/code_provider/seal.py
@@ -1,7 +1,9 @@
 """mse_cli.home.command.code_provider.seal module."""
 
+import sys
 from pathlib import Path
 
+from intel_sgx_ra.quote import Quote
 from intel_sgx_ra.ratls import ratls_verify
 from mse_lib_crypto.seal_box import seal
 
@@ -11,29 +13,42 @@ from mse_cli.log import LOGGER as LOG
 def add_subparser(subparsers):
     """Define the subcommand."""
     parser = subparsers.add_parser(
-        "seal", help="seal the secrets to be share with an MSE app"
+        "seal",
+        help="seal file using NaCl's Seal Box."
+        "Recipient is either raw X25519 public key or "
+        "extracted from RA-TLS certificate with enclave's "
+        "public key in REPORT_DATA field of SGX quote",
     )
 
     parser.add_argument(
-        "--secrets",
-        type=Path,
-        required=True,
-        help="secret file to seal",
-    )
-
-    parser.add_argument(
-        "--cert",
-        required=True,
+        "--input",
         type=Path,
         metavar="FILE",
-        help="path to the ratls certificate",
+        required=True,
+        help="path to the file to seal",
+    )
+
+    group = parser.add_mutually_exclusive_group(required=True)
+
+    group.add_argument(
+        "--receiver-enclave",
+        type=Path,
+        metavar="FILE",
+        help="path to RA-TLS certificate of the enclave",
+    )
+
+    group.add_argument(
+        "--receiver-public-key",
+        type=Path,
+        metavar="FILE",
+        help="path to raw X25519 public key",
     )
 
     parser.add_argument(
         "--output",
         type=Path,
-        required=True,
-        help="directory to write the sealed secrets file",
+        metavar="FILE",
+        help="path to write the file sealed",
     )
 
     parser.set_defaults(func=run)
@@ -41,12 +56,23 @@ def add_subparser(subparsers):
 
 def run(args) -> None:
     """Run the subcommand."""
-    quote = ratls_verify(args.cert)
+    LOG.info("Sealing %s...", args.input)
 
-    enclave_pk = quote.report_body.report_data[32:64]
-    sealed_secrets = seal(args.secrets.read_bytes(), enclave_pk)
+    enclave_pk: bytes
+    if args.receiver_enclave:
+        quote: Quote = ratls_verify(args.receiver_enclave)
+        enclave_pk = quote.report_body.report_data[32:64]
+    else:
+        enclave_pk = args.receiver_public_key.read_bytes()
 
-    sealed_secrets_path: Path = args.output / (args.secrets.name + ".sealed")
-    sealed_secrets_path.write_bytes(sealed_secrets)
+    encrypted_data: bytes = seal(args.input.read_bytes(), enclave_pk)
 
-    LOG.info("Your sealed secrets has been saved at: %s", sealed_secrets_path)
+    if args.output:
+        args.output.write_bytes(encrypted_data)
+        LOG.info("File sealed to %s", args.output)
+    else:
+        LOG.info("Data sucessfully sealed!")
+        LOG.info(
+            "----------------------------------------------------------------------"
+        )
+        sys.stdout.buffer.write(encrypted_data)